+++ src/sock.c
@@ -538,7 +538,7 @@ static void connect_conn(struct mg_connection *c) {
 
 void mg_mgr_poll(struct mg_mgr *mgr, int ms) {
   struct mg_connection *c, *tmp;
-  unsigned long now;
+  unsigned long long now;
 
   mg_iotest(mgr, ms);
   now = mg_millis();
