+++ src/util.h
@@ -20,7 +20,7 @@ int mg_asprintf(char **buf, size_t size, const char *fmt, ...);
 int mg_vasprintf(char **buf, size_t size, const char *fmt, va_list ap);
 int64_t mg_to64(struct mg_str str);
 double mg_time(void);
-unsigned long mg_millis(void);
+unsigned long long mg_millis(void);
 void mg_usleep(unsigned long usecs);
 
 #if MG_ENABLE_FS
